% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-grid-plot.R
\name{frequency_grid_plot}
\alias{frequency_grid_plot}
\title{Frequency grid ggplot}
\usage{
frequency_grid_plot(
  x,
  show_line_grid = FALSE,
  show_line_mode = FALSE,
  label_hypothetical = "NA",
  color_label_hypothetical = "red2",
  color_hypothetical = "red2",
  color_non_hypothetical = "blue2",
  alpha_hypothetical = 1,
  alpha_non_hypothetical = 0.75,
  size_label_hypothetical = 3,
  size_hypothetical = 10,
  size_non_hypothetical = 10,
  shape_hypothetical = 1,
  shape_non_hypothetical = 15,
  expand = 0.1
)
}
\arguments{
\item{x}{A vector with frequencies to visualize.}

\item{show_line_grid}{Logical. Should gridlines be present, crossing at each
observation? Default is \code{FALSE}.}

\item{show_line_mode}{Logical. Should a dashed line demarcate the mode(s)
among known values from the missing values that might add to these modes,
if there are any? Default is \code{FALSE}.}

\item{label_hypothetical}{String. Label used for missing values. Default is
\code{"NA"}.}

\item{color_label_hypothetical, color_hypothetical, color_non_hypothetical}{String. Colors of
the data points. Defaults are \code{"red2"} for missing data points as well as
their labels, and \code{"blue2"} for non-missing data points.}

\item{alpha_hypothetical, alpha_non_hypothetical}{Numeric. Opacity of the data points.
Defaults are \code{1} and \code{0.75}, respectively.}

\item{size_label_hypothetical, size_hypothetical, size_non_hypothetical}{Numeric. Sizes of the
data points. Defaults are \code{3} for the label and \code{10} for both symbols.}

\item{shape_hypothetical, shape_non_hypothetical}{Numeric or string. Signifiers for the
shapes of the data points. Defaults are \code{1} (circle) and \code{15} (square
filled), respectively.}

\item{expand}{Numeric. Padding whitespace between the axes and the data
points. The distance is the same on all four sides due to the grid
structure. Default is \code{0.1}.}
}
\value{
A ggplot object. To save it, call \code{ggplot2::ggsave()}.
}
\description{
NOTE: This function is currently experimental and shouldn't be
relied upon.

Call \code{frequency_grid_plot()} to visualize the absolute frequencies of
values in a vector. Each observation is plotted distinctly, resulting in a
hybrid of a histogram and a scatterplot.
\itemize{
\item Boxes are known values.
\item Circles with \code{NA} labels are missing values.
\item Empty circles are no values at all: They signify that certain unique
values would have to be more frequent in order for all unique values to be
equally frequent.
}
}
\section{Limitations}{
 Certain assumptions about missing values are currently
hard-coded in the function. In the future, they should become optional.
These assumptions are:
\itemize{
\item All missings represent a known value. For example, in \code{c(1, 2, NA)}, the
\code{NA} is either \code{1} or \code{2}.
\item The missings are as evenly distributed across known values as possible.
Therefore, in \code{c(1, 2, NA, NA)}, one \code{NA} is a \code{1} and the other one is a
\code{2}. This is clearly not reasonable as a general assumption. It is derived
from moder's way of determining possible extreme cases.
}
}

\examples{
x <- c("a", "a", "a", "b", "b", "c", NA, NA, NA, NA, NA)

# Basic usage:
frequency_grid_plot(x)

# With "N/A" as a marker of missing values
# instead of "NA":
frequency_grid_plot(x, label_hypothetical = "N/A")

# Black and white mode:
frequency_grid_plot(
  x, color_label_hypothetical = "black",
  color_hypothetical = "black", color_non_hypothetical = "black"
)
}
\seealso{
\code{\link[=frequency_grid_df]{frequency_grid_df()}}, which forms the basis of the current
function.
}
