% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is-unimodal-multimodal}
\alias{is-unimodal-multimodal}
\alias{is_unimodal}
\alias{is_multimodal}
\title{Is a vector unimodal or multimodal?}
\usage{
is_unimodal(x, na.rm = FALSE, max_unique = NULL)

is_multimodal(x, na.rm = FALSE, max_unique = NULL)
}
\arguments{
\item{x}{A vector to search for its modes.}

\item{na.rm}{Logical. Should missing values in \code{x} be removed, i.e., ignored?
Default is \code{FALSE}.}

\item{max_unique}{Numeric or string. If the maximum number of unique values
in \code{x} is known, set \code{max_unique} to that number. This puts a limit on how
many unique values may be represented by \code{NA}s. Set it to \code{"known"} instead
if no values beyond those already known can occur in \code{x}, or if \code{x} is a
factor (it will then count the levels). Default is \code{NULL}, which assumes no
maximum.}
}
\value{
Logical (length 1).
}
\description{
Two complementary functions:
\itemize{
\item \code{is_unimodal()} checks whether a vector has exactly one mode.
\item \code{is_multimodal()} checks whether a vector has multiple modes.
}
}
\details{
These functions are thin wrappers around \code{\link[=mode_count]{mode_count()}}, a more
general tool. See examples there for \code{NA} handling.
}
\section{Factors and \code{max_unique}}{
 If \code{x} is a factor, \code{max_unique} should be
\code{"known"} or there is a warning. This is because a factor's levels are
supposed to include all of its possible values.
}

\examples{
# Only one mode:
is_unimodal(c(5, 5, 6))
is_multimodal(c(5, 5, 6))

# Two modes:
is_unimodal(c(5, 5, 6, 6))
is_multimodal(c(5, 5, 6, 6))

# Unclear because of the missing value:
is_unimodal(c(5, 5, 6, NA))
is_multimodal(c(5, 5, 6, NA))
}
\seealso{
\code{\link[=is_uniform]{is_uniform()}} to check for a uniform distribution.
}
