% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_uniform}
\alias{is_uniform}
\title{Are the values uniformly distributed?}
\usage{
is_uniform(x, na.rm = FALSE, max_unique = NULL)
}
\arguments{
\item{x}{A vector to search for its modes.}

\item{na.rm}{Logical. Should missing values in \code{x} be removed, i.e., ignored?
Default is \code{FALSE}.}

\item{max_unique}{Numeric or string. If the maximum number of unique values
in \code{x} is known, set \code{max_unique} to that number. This puts a limit on how
many unique values may be represented by \code{NA}s. Set it to \code{"known"} instead
if no values beyond those already known can occur in \code{x}, or if \code{x} is a
factor (it will then count the levels). Default is \code{NULL}, which assumes no
maximum.}
}
\value{
Logical (length 1).
}
\description{
\code{is_uniform()} checks whether all values in a given vector are
equally frequent, i.e., uniformly distributed. The mode is trivial in such
cases.
}
\details{
The function returns \code{TRUE} if \code{x} has length < 3 because no value
is more frequent than another one. Otherwise, it returns \code{NA} in these
cases:
\itemize{
\item Some \code{x} values are missing, all known values are equal, and \code{max_unique}
is not \code{1} or \code{"known"}. Thus, it is unknown whether there are values with
different frequencies.
\item All known values are modes if the \code{NA}s "fill up" all non-modal known
values exactly, i.e., without any \code{NA}s remaining. These values might then
hypothetically be modes.
\item There are so many missing values that they might form mode-sized groups
of values that are not among the known values, and the number of \code{NA}s is
divisible by the modal frequency so that all (partly hypothetical) values
might be equally frequent. You can limit the number of such hypothetical
values by specifying \code{max_unique}. The function might then return \code{FALSE}
instead of \code{NA}.
\item Some \code{NA}s remain after "filling up" the non-modal known values with
\code{NA}s and / or hypothetically forming new mode-sized groups, and the number
of remaining \code{NA}s is divisible by the number of unique known values. Thus,
these surplus \code{NA}s can be evenly distributed across the known (or,
perhaps, partly hypothesized) values. Thus, a uniform distribution is still
possible. \code{max_unique} can be relevant here, as well.
}
}
\section{Factors and \code{max_unique}}{
 If \code{x} is a factor, \code{max_unique} should be
\code{"known"} or there is a warning. This is because a factor's levels are
supposed to include all of its possible values.
}

\examples{
# The distribution is uniform
# if all values are equal...
is_uniform(c(1, 1, 1))

# ...and even if there are multiple unique
# values but they are all equally frequent:
is_uniform(c(1, 1, 2, 2))

# This includes cases where
# all values are different:
is_uniform(c(1, 2, 3))

# Here, the mode is nontrivial
# because `1` is more frequent than `2`:
is_uniform(c(1, 1, 2))

# Two of the `NA`s might be `8`s, and
# the other three might represent a value
# different from both `7` and `8`. Thus,
# it's possible that all three distinct
# values are equally frequent:
is_uniform(c(7, 7, 7, 8, rep(NA, 5)))

# The same is not true if all values,
# even the missing ones, must represent
# one of the known values:
is_uniform(c(7, 7, 7, 8, rep(NA, 5)), max_unique = "known")
}
