% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode-count-possible}
\alias{mode-count-possible}
\alias{mode_count_possible_min}
\alias{mode_count_possible_max}
\title{Count possible modes}
\usage{
mode_count_possible_min(x, accept_known = TRUE)

mode_count_possible_max(x)
}
\arguments{
\item{x}{A vector to search for its possible modes.}

\item{accept_known}{Boolean. Only in \code{mode_possible_min()}. If set to
\code{FALSE}, the function returns \code{NA} if the set of all modes can't be
determined due to missing values. Default is \code{TRUE}.}
}
\value{
Integer. Minimal and maximal number of modes (values tied for most
frequent) in \code{x}. If the functions can't determine these possible
modes because of missing values, they return \code{NA} instead.
}
\description{
\code{mode_count_possible_min()} and \code{mode_count_possible_max()}
determine the minimal and maximal number of modes from among known modes,
given the number of missing values.

They are wrappers around \code{\link[=mode_possible_min]{mode_possible_min()}} and \code{\link[=mode_possible_max]{mode_possible_max()}}.
}
\seealso{
The examples for \code{\link[=mode_possible_min]{mode_possible_min()}} and \code{\link[=mode_possible_max]{mode_possible_max()}}, to
make it more clear how the present functions work.
}
