% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode-possible-count}
\alias{mode-possible-count}
\alias{mode_count_possible_min}
\alias{mode_count_possible_max}
\title{Count possible modes}
\usage{
mode_count_possible_min(x)

mode_count_possible_max(x)
}
\arguments{
\item{x}{A vector to search for its possible modes.}
}
\value{
Integer. Minimal and maximal number of modes (values tied for most
frequent) in \code{x}. If \code{mode_count_possible_min()} can't determine these
possible modes because of missing values, it returns \code{NA} instead.
}
\description{
\code{mode_count_possible_min()} and \code{mode_count_possible_max()} determine the
minimal and maximal number of modes, given the number of missing values.
}
\seealso{
\code{\link[=mode_possible_min]{mode_possible_min()}} and \code{\link[=mode_possible_max]{mode_possible_max()}} for determining the
sets of these possible modes. The examples there will make it clear how the
present functions work.
}
