% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode-df.R
\name{mode_df}
\alias{mode_df}
\title{Tabulate mode estimates with the certainty about them}
\usage{
mode_df(
  x,
  method = c("first", "all", "single"),
  na.rm.from = c("first", "last", "random"),
  accept = FALSE,
  multiple = c("NA", "min", "max", "mean", "median", "first", "last", "random")
)
}
\arguments{
\item{x}{List of vectors. Each vector needs to be numeric or similar. Note
that data frames are lists, so \code{x} can be a data frame.}

\item{method}{String. How to determine the mode(s)? Options are:
\itemize{
\item \code{"first"} for \code{\link[=mode_first]{mode_first()}}, the default.
\item \code{"all"} for \code{\link[=mode_all]{mode_all()}}. This may return multiple values per estimate.
\item \code{"single"} for \code{\link[=mode_single]{mode_single()}}. The only option that can return \code{NA}
estimates.
}}

\item{na.rm.from}{String. Only relevant to the default \code{method = "first"}.
Where to start when removing \code{NA}s from \code{x}? Options are \code{"start"},
\code{"end"}, and \code{"random"}. Default is \code{"start"}.}

\item{accept}{Passed on to \code{\link[=mode_first]{mode_first()}} and \code{\link[=mode_single]{mode_single()}}. Default is
\code{FALSE}.}

\item{multiple}{Passed on to \code{\link[=mode_single]{mode_single()}}. Default is \code{"NA"}.}
}
\value{
Tibble (data frame) with these columns:
\itemize{
\item \code{term}: the names of \code{x} elements. Only present if any are named.
\item \code{estimate}: the modes of \code{x} elements, ignoring as many \code{NA}s as
necessary. List-column if \code{method = "all"}.
\item \code{certainty}: \code{TRUE} if the corresponding estimate is certain to be the
true mode, and \code{FALSE} if this is unclear due to missing values.
\item \code{na_ignored}: the number of missing values that had to be ignored to
arrive at the estimate.
\item \code{na_total}: the total number of missing values.
\item \code{rate_ignored_na}: the proportion of missing values that had to be
ignored from among all missing values.
\item \code{sum_total}: the total number of values, missing or not.
\item \code{rate_ignored_sum}: the proportion of missing values that had to be
ignored from among all values, missing or not.
}
}
\description{
\code{mode_df()} takes a data frame (or another list) of numeric
vectors and computes the mode or modes of each element. Where the true mode
is unknown due to missing values, more and more \code{NA}s are ignored until an
estimate for the mode is found.

Estimates are presented along with information about whether they are known
to be the true mode, how many \code{NA}s had to be ignored during estimation,
the rate of ignored \code{NA}s, etc.
}
\details{
The function deals with missing values (\code{NA}s) by first checking
whether they make the true mode unknown. If they do, it removes one \code{NA},
then checks again; and so on until an estimate is found.

This strategy is based on the \code{na.rm.amount} argument of \code{\link[=mode_first]{mode_first()}},
\code{\link[=mode_all]{mode_all()}}, and \code{\link[=mode_single]{mode_single()}}. It represents a middle way between
simply ignoring all \code{NA}s and not even trying to compute an estimate.
Instead, it only removes the minimum number of \code{NA}s necessary, because
some distributions have a known mode (or set of modes) even if some of
their values are missing. By keeping track of the removed \code{NA}s,
\code{mode_df()} quantifies the uncertainty about its estimates.
}
\examples{
# Use a list of numeric vectors:
my_list <- list(
  a = 1:15,
  b = c(1, 1, NA),
  c = c(4, 4, NA, NA, NA, NA),
  d = c(96, 24, 3, NA)
)

mode_df(my_list)

# Data frames are allowed:
mode_df(iris[1:4])
}
