% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{mode_frequency}
\alias{mode_frequency}
\title{Modal frequency}
\usage{
mode_frequency(x, na.rm = FALSE, max_unique = NULL)
}
\arguments{
\item{x}{A vector to check for its modal frequency.}

\item{na.rm}{Logical. Should missing values in \code{x} be removed, i.e., ignored?
Default is \code{FALSE}.}

\item{max_unique}{Numeric or string. If the maximum number of unique values
in \code{x} is known, set \code{max_unique} to that number. This puts a limit on how
many unique values may be represented by \code{NA}s. Set it to \code{"known"} instead
if no values beyond those already known can occur in \code{x}, or if \code{x} is a
factor (it will then count the levels). Default is \code{NULL}, which assumes no
maximum.}
}
\value{
Integer (length 1).
}
\description{
Call \code{mode_frequency()} to get the number of times that a
vector's mode appears in the vector.

See \code{\link[=mode_frequency_range]{mode_frequency_range()}} for bounds on an unknown frequency.
}
\details{
By default, the function returns \code{NA} if any values are missing.
That is because missings make the frequency uncertain even if the mode is
known: any missing value may or may not be an instance of the mode, and
hence count towards the modal frequency.
}
\examples{
# The mode, `9`, appears three times:
mode_frequency(c(7, 8, 8, 9, 9, 9))

# Missing values make the frequency
# unknown, even if the mode is obvious:
mode_frequency(c(1, 1, NA))

# You can ignore this problem and
# determine the frequency among known values
# (there should be good reasons for this!):
mode_frequency(c(1, 1, NA), na.rm = TRUE)

# In a multimodal distribution, the modal
# frequency is that of a single mode:
mode_frequency(c(3, 3, 4, 4, 5, 5, 6))
}
\seealso{
\code{\link[=mode_first]{mode_first()}}, which the function wraps.
}
