% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{mode_is_trivial}
\alias{mode_is_trivial}
\title{Is the mode trivial?}
\usage{
mode_is_trivial(x, na.rm = FALSE, max_unique = NULL)
}
\arguments{
\item{x}{A vector to search for its modes.}

\item{na.rm}{Logical. Should missing values in \code{x} be removed, i.e., ignored?
Default is \code{FALSE}.}

\item{max_unique}{Numeric or string. If the maximum number of unique values
in \code{x} is known, set \code{max_unique} to that number. This puts a limit on how
many unique values may be represented by \code{NA}s. Set it to \code{"known"} instead
if no values beyond those already known can occur in \code{x}, or if \code{x} is a
factor (it will then count the levels). Default is \code{NULL}, which assumes no
maximum.}
}
\value{
Logical (length 1).
}
\description{
\code{mode_is_trivial()} was renamed to \code{\link[=is_uniform]{is_uniform()}}. Please use
this new function instead.

The function with the old name is deprecated and will be removed in a
future version. It is retained for now, but using it will cause a warning.
}
