% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-grid-df.R
\name{frequency_grid_df}
\alias{frequency_grid_df}
\title{Frequency grid data frame}
\usage{
frequency_grid_df(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A data frame with these columns:
\itemize{
\item \code{x}: The input vector, with each unique known value repeated to be as
frequent as the most frequent one.
\item \code{freq} (integer): Hypothetical frequency of each \code{x} value.
\item \code{is_missing} (Boolean): Is the observation absent from the input vector?
\item \code{can_be_filled} (Boolean): Are there enough \code{NA}s so that one of them might
hypothetically represent the \code{x} value in question, implying that there
would be at least as many observations of that value as the respective
frequency (\code{freq}) indicates?
\item \code{is_supermodal} (Boolean): Is the frequency of this value greater than the
maximum frequency among known values?
}
}
\description{
\code{frequency_grid_df()} takes a vector and creates an extended frequency table
about it. Internally, this is used as a basis for \code{frequency_grid_plot()}.
}
\section{Limitations}{
 See the limitations section of \code{frequency_grid_plot()}.
}

\examples{
x <- c("a", "a", "a", "b", "b", "c", NA, NA, NA, NA, NA)
frequency_grid_df(x)
}
