% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode-possible}
\alias{mode-possible}
\alias{mode_possible_min}
\alias{mode_possible_max}
\title{Possible modes}
\usage{
mode_possible_min(x)

mode_possible_max(x)
}
\arguments{
\item{x}{A vector to search for its possible modes.}
}
\value{
A vector with the minimal or maximal possible modes (values tied for
most frequent) in \code{x}. If the functions can't determine these possible
modes because of missing values, they return \code{NA} instead.
}
\description{
\code{mode_possible_min()} and \code{mode_possible_max()} determine the minimal and
maximal sets of modes from among known modes, given the number of missing
values.
}
\examples{
# "a" is guaranteed to be a mode,
# "b" might also be one, but "c" is impossible:
mode_possible_min(c("a", "a", "a", "b", "b", "c", NA))
mode_possible_max(c("a", "a", "a", "b", "b", "c", NA))

# Only `7` can possibly be the mode
# because, even if `NA` is `8`, it's
# still less frequent than `7`:
mode_possible_min(c(7, 7, 7, 7, 7, 8, 8, NA))
mode_possible_max(c(7, 7, 7, 7, 7, 8, 8, NA))

# No clear minimum of modes (because
# `NA` may tip the balance towards a
# single mode), but a clear maximum:
mode_possible_min(c(1, 1, 2, 2, NA))
mode_possible_max(c(1, 1, 2, 2, NA))
}
\seealso{
\code{\link[=mode_count_possible_min]{mode_count_possible_min()}} and \code{\link[=mode_count_possible_max]{mode_count_possible_max()}} for
counting these possible modes.
}
