% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode_all}
\alias{mode_all}
\title{All modes}
\usage{
mode_all(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector to search for its modes.}

\item{na.rm}{Boolean. Should missing values in \code{x} be removed before
computation proceeds? Default is \code{FALSE}.}
}
\value{
A vector with all modes (values tied for most frequent) from \code{x}. If
the modes can't be determined because of missing values, returns \code{NA}
instead.
}
\description{
\code{mode_all()} returns the set of all modes in a vector.
}
\seealso{
\itemize{
\item \code{\link[=mode_first]{mode_first()}} for the first-appearing mode.
\item \code{\link[=mode_single]{mode_single()}} for the \emph{only} mode, or \code{NA} if there are more.
}
}
