% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode_count}
\alias{mode_count}
\title{Number of modes}
\usage{
mode_count(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector to search for its modes.}

\item{na.rm}{Boolean. Should missing values in \code{x} be removed before
computation proceeds? Default is \code{FALSE}.}
}
\value{
Integer. Number of modes in \code{x}. If the modes can't be determined
because of missing values, returns \code{NA} instead.
}
\description{
\code{mode_count()} counts the modes in a vector. Thin wrapper around
\code{\link[=mode_all]{mode_all()}}.
}
