% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode_count_range}
\alias{mode_count_range}
\title{Modal count range}
\usage{
mode_count_range(x)
}
\arguments{
\item{x}{A vector to search for its possible modes.}
}
\value{
Integer (length 2). Minimal and maximal number of modes (values tied
for most frequent) in \code{x}. If these possible modescan't be determined
because of missing values, the function returns \code{NA} instead.
}
\description{
\code{mode_count_range()} determines the minimal and maximal number
of modes from among known modes, given the number of missing values.
}
\seealso{
The examples for \code{\link[=mode_possible_min]{mode_possible_min()}} and \code{\link[=mode_possible_max]{mode_possible_max()}}, to
make it more clear how the present function works. (It wraps these two.)
}
