% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode_count_range}
\alias{mode_count_range}
\title{Modal count range}
\usage{
mode_count_range(x)
}
\arguments{
\item{x}{A vector to search for its possible modes.}
}
\value{
Integer (length 2). Minimal and maximal number of modes (values tied
for most frequent) in \code{x}. If these possible modes cannot be determined
because of missing values, the function returns \code{NA} instead.
}
\description{
\code{mode_count_range()} determines the minimal and maximal number
of modes from among known modes, given the number of missing values.
}
\examples{
# If `NA` is `7` or `8`, that number is
# the only mode; otherwise, both numbers
# are modes:
mode_count_range(c(7, 7, 8, 8, NA))

# Same result here -- `7` is the only mode
# unless `NA` is secretly `8`, in which case
# there are two modes:
mode_count_range(c(7, 7, 7, 8, 8, NA))

# But now, there is now way for `8` to be
# as frequent as `7`:
mode_count_range(c(7, 7, 7, 7, 8, 8, NA))
}
\seealso{
\code{\link[=mode_possible_min]{mode_possible_min()}} and \code{\link[=mode_possible_max]{mode_possible_max()}}, which the present
function wraps.
}
