% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode_count_range}
\alias{mode_count_range}
\title{Modal count range}
\usage{
mode_count_range(x, exclusive = FALSE, max_unique = NULL)
}
\arguments{
\item{x}{A vector to search for its possible modes.}

\item{exclusive}{Boolean. Can \code{NA}s only represent known values? Set
\code{exclusive} to \code{TRUE} if you are certain that there are no other values
behind the \code{NA}s, such that each \code{NA} masks one of the known values.
Default is \code{FALSE}.}

\item{max_unique}{Integer. Maximum number of distinct values in \code{x}. Default
is \code{NULL}, which assumes no maximum.}
}
\value{
Integer (length 2). Minimal and maximal number of modes (values tied
for most frequent) in \code{x}.
}
\description{
\code{mode_count_range()} determines the minimal and maximal number
of modes given the number of missing values.
}
\details{
If \code{x} is a factor, \code{max_unique} is automatically set to the number
of levels in \code{x}. If the user sets \code{max_unique} to another value, it isn't
reset automatically but there is a warning. This is because a factor's
levels are supposed to be all of its possible values. For the same reason,
there is a warning if \code{x} is a factor and \code{exclusive} is \code{FALSE}.
}
\examples{
# If `NA` is `7` or `8`, that number is
# the only mode; otherwise, both numbers
# are modes:
mode_count_range(c(7, 7, 8, 8, NA))

# Same result here -- `7` is the only mode
# unless `NA` is secretly `8`, in which case
# there are two modes:
mode_count_range(c(7, 7, 7, 8, 8, NA))

# But now, there is now way for `8` to be
# as frequent as `7`:
mode_count_range(c(7, 7, 7, 7, 8, 8, NA))

# The `NA`s might form a new mode here
# if they are both, e.g., `9`:
mode_count_range(c(7, 7, 8, 8, NA, NA))

# However, if there can be no values beyond
# those already known -- `7` and `8` --
# the `NA`s can't form a new mode.
# Specify this with `exclusive = TRUE`:
mode_count_range(c(7, 7, 8, 8, NA, NA), exclusive = TRUE)
}
