% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode_first}
\alias{mode_first}
\title{The first-appearing mode}
\usage{
mode_first(x, na.rm = FALSE, first_known = TRUE)
}
\arguments{
\item{x}{A vector to search for its first mode.}

\item{na.rm}{Boolean. Should missing values in \code{x} be removed before
computation proceeds? Default is \code{FALSE}.}

\item{first_known}{Boolean. Should the first-appearing value known to be a
mode be accepted? If \code{FALSE}, returns \code{NA} if a value that appears earlier
might be another mode due to missing values. Default is \code{TRUE}.}
}
\value{
The first mode (most frequent value) in \code{x}. If it can't be
determined because of missing values, returns \code{NA} instead.
}
\description{
\code{mode_first()} returns the mode that appears first in a vector, i.e., before
any other modes.
}
\seealso{
\itemize{
\item \code{\link[=mode_all]{mode_all()}} for the full set of modes.
\item \code{\link[=mode_single]{mode_single()}} for the \emph{only} mode, or \code{NA} if there are more.
}
}
