% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode_frequency}
\alias{mode_frequency}
\title{Modal frequency}
\usage{
mode_frequency(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector to search for its first mode.}

\item{na.rm}{Boolean. Should missing values in \code{x} be removed before
computation proceeds? Default is \code{FALSE}.}
}
\value{
Integer (length 1) or \code{NA}.
}
\description{
Call \code{mode_frequency()} to get the number of times that a
vector's mode appears in the vector.

By default, the function returns \code{NA} if any values are missing. That is
because missings make the frequency uncertain even if the mode is known.
}
\seealso{
\code{\link[=mode_first]{mode_first()}}, which the function wraps.
}
