% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode_frequency_range}
\alias{mode_frequency_range}
\title{Modal frequency range}
\usage{
mode_frequency_range(x)
}
\arguments{
\item{x}{A vector to search for its first mode.}
}
\value{
Integer (length 2) or \code{NA}.
}
\description{
Call \code{mode_frequency_range()} to get the number of times that a
vector's mode appears in the vector. It returns the range of the frequency
in a length-2 vector. The lower bound assumes that no \code{NA}s count towards
the mode; the upper bound assumes that all \code{NA}s do. If there are no \code{NA}s,
the bounds are identical.

If the mode can't be determined, \code{NA} is returned instead.
}
\examples{
# The mode is `7`. It appears four or
# five times because the `NA` might
mode_frequency_range(c(7, 7, 7, 7, 8, 8, NA))

# `"c"`, `"d"`, and `"e"` are the modes,
# and each of them appears twice:
mode_frequency_range(c("a", "b", "c", "c", "d", "d", "e", "e"))
}
\seealso{
\itemize{
\item \code{\link[=mode_first]{mode_first()}}, which the function wraps.
\item \code{\link[=mode_frequency]{mode_frequency()}}, for the precise frequency (or \code{NA} if it can't be
determined).
}
}
